+++ packages/react/shared/src/WebtrekkSmartPixelReact.js
@@ -108,6 +108,15 @@ class WebtrekkSmartPixelReact {
     }
 
     /**
+     * @param {object} data
+     */
+    engage(data = emptyObject) {
+        this.call(function(pix) {
+            pix.engage.data.add(data);
+        });
+    }
+
+    /**
      * @param {string} id
      * @param {object} data
      * @param {boolean} validation