+++ packages/react/shared/src/components/data/WebtrekkCustomerData.js
@@ -26,6 +26,14 @@ WebtrekkCustomerData.propTypes = {
     street: PropTypes.string,
     streetNumber: PropTypes.string,
     validation: PropTypes.bool,
+    registrationEmail: PropTypes.string,
+    registrationGroupId: PropTypes.string,
+    registrationMode: PropTypes.string,
+    registrationFirstName: PropTypes.string,
+    registrationLastName: PropTypes.string,
+    registrationGender: PropTypes.string,
+    registrationTitle: PropTypes.string,
+    registrationOptin: PropTypes.bool,
     category: PropTypes.objectOf(PropTypes.string),
     sendInstantly: PropTypes.bool
 };
@@ -46,6 +54,14 @@ WebtrekkCustomerData.defaultProps = {
     street: null,
     streetNumber: null,
     validation: null,
+    registrationEmail: null,
+    registrationGroupId: null,
+    registrationMode: null,
+    registrationFirstName: null,
+    registrationLastName: null,
+    registrationGender: null,
+    registrationTitle: null,
+    registrationOptin: null,
     category: null,
     sendInstantly: false
 };